feat: add demo to cli
