+++ package/demo/json/en.json
@@ -1,0 +1,6 @@
+{
+  "home": {
+    "title/main": "Replexica",
+    "description/dev": "Replexica is an AI-powered localization-as-a-service platform for modern SaaS."
+  }
+}
