+++ package/demo/yaml-root-key/en.yml
@@ -1,0 +1,4 @@
+en:
+  home:
+    title/main: "Replexica"
+    description/dev: "Replexica is an AI-powered localization-as-a-service platform for modern SaaS."
