+++ package/demo/yaml/en.yml
@@ -1,0 +1,3 @@
+home:
+  title/main: "Replexica"
+  description/dev: "Replexica is an AI-powered localization-as-a-service platform for modern SaaS."
