+++ packages/cli/src/index.ts
@@ -4,17 +4,15 @@ dotenv.config();
 
 import { Command } from 'commander';
 
-import i18nCmd from './i18n';
 import authCmd from './auth';
 import initCmd from './init';
 import configCmd from './config';
-import i18nNew from './i18n.new';
+import i18nCmd from './i18n';
 
 export default new Command()
   .name('replexica')
   .description('Replexica CLI')
   .helpOption('-h, --help', 'Show help')
-  .addCommand(i18nNew)
   .addCommand(i18nCmd)
   .addCommand(authCmd)
   .addCommand(initCmd)
